feat: remove page-based discussion logic
