+++ web/app/routes.js
@@ -5,7 +5,5 @@ export default [
   route("question", "routes/question.jsx"),
   route("article", "routes/article.jsx"),
   route("articles/:id", "routes/article.$id.jsx"),
-  route("articles/:id/questions", "routes/article.$id.questions.jsx"),
-  route("discussions", "routes/discussions.jsx"),
-  route("discussions/results", "pages/DiscussionResults.jsx")
+  route("articles/:id/questions", "routes/article.$id.questions.jsx")
 ];
